pyproject.toml
[tool.poetry]
name = "wollala-upbit"
version = "0.1.0"
description = ""
authors = ["Bumjoon Park <bumjoon.park@qt.io>"]

[tool.poetry.dependencies]
python = "=3.8"
PySide6 = "^6.2.2"
requests = "^2.26.0"
pandas = "^1.3.5"
upbit-client = "^1.2.2"
pyinstaller = "^4.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
